CCTower Slow&Stun&DEcrease処理

+++ Assets/Scripts/Enemy/Monster.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Monster : Enemy {
-    public Monster(string name, int lv, int hp, int speed) {
+    public Monster(string name, int lv, int hp, float speed) {
         Name = name;
         Lv = lv;
         Hp = hp;
